Move application class and update packages
