+++ HackerPlatform-Backend/src/test/java/com/myorg/hackerplatform/HackerPlatformBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.myorg.hackerplatform.HackerPlatform.Backend;
+package com.myorg.hackerplatform;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
